speed up loading workgraph in web
- save short workgraph data in base.extras
- serialize each task, instead of whole workgraph
- only query task data for task detail
